Add starter file for required JavaScript
The provided file leads the student a bit more on how they should
structure their solution. A solution to the first problem is provided
to additional guidance on what a solution might look like, and
a function stub for the next problem is provide to lead the student
into a next step.